improve test coverage
